// src/utility/email-provider/email-category.enum.ts
export enum EmailCategory {
  ATTENDANCE_CHECKIN = 'ATTENDANCE_CHECKIN',
  BIRTHDAY = 'BIRTHDAY',
  EVENT_REMINDER = 'EVENT_REMINDER',
  PRAYER_REMINDER = 'PRAYER_REMINDER',
  FOLLOW_UP = 'FOLLOW_UP',
  ASSET_ALERTS = 'ASSET_ALERTS',
  GIVING_RECEIPT = 'GIVING_RECEIPT',
  FINANCE_ALERTS = 'FINANCE_ALERTS',
  SESSION_REPORT = 'SESSION_REPORT',
  INCIDENT_REPORT = 'INCIDENT_REPORT',
  CHILDREN_CHURCH = 'CHILDREN_CHURCH',
  LOGIN_ALERT = 'LOGIN_ALERT',
  SERVICE_PROGRAMME_ASSIGNMENT = 'SERVICE_PROGRAMME_ASSIGNMENT',
  PASTOR_FEEDBACK = 'PASTOR_FEEDBACK',
  MEMBERSHIP_ANNIVERSARY = 'MEMBERSHIP_ANNIVERSARY',
  ASSIGNMENT_REMINDER = 'ASSIGNMENT_REMINDER',
  CLASS_SESSION_REMINDER = 'CLASS_SESSION_REMINDER',
}